requirements.txt
zenml==0.84.2
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.25.0
pymongo==4.10.1
python-dotenv==1.0.1
pandas==2.2.3
pydantic==2.10.3
python-dateutil==2.9.0